Added main menu
This is setup for changing the controls. There are two menus. One is main menu while the other is options menu. Both work fine. Not much needs to be done for the controls besides setting up to have it save changes to the ini file and update the global variable associated with that input.