FIXED IT. Controls now work and swap
The issue was that mpos would become non zero and therefore wouldn't run the script that actually changed the controls because it was in a switch statement that require mpos to be 0. In other words, changing either up or down would change the mpos but left, right, and action were fine.